+++ src/features/schedule/ui/DescriptionSection.tsx
