+++ src/features/schedule/ui/LocationSection.tsx
