+++ src/features/schedule/ui/ParticipantsSection.tsx
