+++ src/features/schedule/ui/TimeSection.tsx
