+++ src/features/schedule/ui/TitleSection.tsx
