+++ src/features/schedule/ui/index.ts
