+++ src/pages/teamroom/create/index.tsx
@@ -1,16 +1,16 @@
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import {
-  formatDateString,
-  formatMonthDayLabel,
-} from '@/entities/calendar/lib/recurrence';
 import { useCreateTeamRoom } from '@/entities/teamroom/hooks/useTeamRoom';
 import { useTeamRoomCreateStore } from '@/entities/teamroom/model/teamroom-create-store';
 import {
   DEFAULT_TEAMROOM_IMAGE_ID,
   TEAMROOM_IMAGES,
 } from '@/shared/constants/teamroom-images';
+import {
+  formatDateString,
+  formatMonthDayLabel,
+} from '@/shared/lib/date/recurrence';
 import BottomButton from '@/shared/ui/button/BottomButton';
 import TopBar from '@/shared/ui/header/TopBar';
 import { useModalStore } from '@/shared/ui/modal/model/modal-store';
