+++ src/shared/ui/modal/CalendarModal.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 
 import Calendar from '@/shared/ui/Calendar';
+import CalendarHeader from '@/shared/ui/CalendarHeader';
 import ModalDim from '@/shared/ui/modal/ModalDim';
-import CalendarHeader from '@/widgets/calendar/ui/CalendarHeader';
 
 import { useModalStore } from './model/modal-store';
 import { CalendarModalOptions } from './model/types';
